requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
python-dotenv
PyMuPDF
faiss-cpu
tiktoken
jinja2
unstructured[pdf]
pdfminer.six
pi-heif
python-telegram-bot==20.7
gspread
google-auth
google-api-python-client

